Make cache purging generic and re-use for all data used by the frontend
