+++ cms/src/globals/Home.ts
@@ -5,7 +5,7 @@ import { ImageWithFloatingTextBlock } from "../blocks/ImageWithFloatingText";
 import { AccommodationSelectorBlock } from "../blocks/AccommodationSelector";
 import { LeadBlock } from "../blocks/Lead";
 import { FeaturesBlock } from "../blocks/Features";
-import { resolve6 } from "dns/promises";
+import { makeCachePurgeHook } from "../hooks/cachePurgeHook";
 
 export const Home: GlobalConfig = {
   slug: "home",
@@ -15,54 +15,7 @@ export const Home: GlobalConfig = {
   },
   access: { read: () => true },
   hooks: {
-    afterChange: [
-      async function afterChangeHook() {
-        try {
-          if (!process.env.CACHE_PURGE_TARGET_TYPE) {
-            throw new Error("CACHE_PURGE_TARGET is not set");
-          }
-          if (!process.env.CACHE_PURGE_TARGET_ARG) {
-            throw new Error("CACHE_PURGE_TARGET_ARG is not set");
-          }
-
-          switch (process.env.CACHE_PURGE_TARGET_TYPE) {
-            case "single":
-              await purgeCache(process.env.CACHE_PURGE_TARGET_ARG);
-              break;
-            case "fly":
-              const [appName, port] =
-                process.env.CACHE_PURGE_TARGET_ARG.split(",");
-              console.log(
-                `Determining Fly frontend VM URLs for cache purge (app=${appName}, port=${port})`,
-              );
-
-              const urls = await queryFlyVmUrls(appName, parseInt(port, 10));
-
-              console.log(`Purging cache at ${urls.length} frontend VMs`);
-
-              const results = await Promise.allSettled(urls.map(purgeCache));
-              const failed = results.filter(isPromiseRejectedResult);
-
-              if (failed.length === 0) {
-                console.log(
-                  `Successfully purged cache at ${urls.length} frontend VMs`,
-                );
-              } else {
-                console.error(
-                  `Failed to purge cache at ${failed.length} frontend VMs:\n${failed.map((r, i) => `[${i}] ${r}`).join("\n")}`,
-                );
-              }
-              break;
-            default:
-              throw new Error(
-                `Unsupported CACHE_PURGE_TARGET type: ${process.env.CACHE_PURGE_TARGET_TYPE}`,
-              );
-          }
-        } catch (e) {
-          console.error("Failed to purge cache:", e);
-        }
-      },
-    ],
+    afterChange: [makeCachePurgeHook("globals/home")],
   },
   fields: [
     slidesField,
@@ -108,32 +61,3 @@ export const Home: GlobalConfig = {
     },
   ],
 };
-
-async function queryFlyVmUrls(appName: string, port: number) {
-  const address = `global.${appName}.internal`;
-  const ipv6s = await resolve6(address);
-  const urls = ipv6s.map((ip) => `http://[${ip}]:${port}/purge-cache`);
-
-  return urls;
-}
-
-async function purgeCache(targetUrl: string) {
-  console.log(`Purging cache at ${targetUrl} for globals/home`);
-  const response = await fetch(targetUrl, {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({ url: "globals/home" }),
-  });
-
-  if (!response.ok) {
-    throw new Error(`Failed to purge cache at ${targetUrl} for globals/home`);
-  }
-}
-
-function isPromiseRejectedResult(
-  result: PromiseSettledResult<unknown>,
-): result is PromiseRejectedResult {
-  return result.status === "rejected";
-}
